break supplypoint type out into separate file
